// src/redux/characterList/types.ts
export enum Types {
  GET_CHARACTERS = '@CHARACTER_LIST/GET_CHARACTERS',
  GET_CHARACTERS_BY_NAME = '@CHARACTER_LIST/GET_CHARACTERS_BY_NAME',
  FILL_CHARACTERS = '@CHARACTER_LIST/FILL_CHARACTERS',
  FILL_FILTERED_CHARACTERS = '@CHARACTER_LIST/FILL_FILTERED_CHARACTERS',
  LOAD_MORE_FILTERED = '@CHARACTER_LIST/LOAD_MORE_FILTERED',
  FILL_MORE_FILTERED = '@CHARACTER_LIST/FILL_MORE_FILTERED',
  CLEAR_FILTERED = '@CHARACTER_LIST/CLEAR_FILTERED',
  IS_LOADING_MORE = '@CHARACTER_LIST/IS_LOADING_MORE',
}